# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dlavie Chat API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts/feed:
    get:
      operationId: getFeed
      tags: [posts]
      summary: Get social feed
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Feed posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Post"
                  hasMore:
                    type: boolean
                required: [posts, hasMore]
  /posts:
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "200":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/user/{userId}:
    get:
      operationId: getUserPosts
      tags: [posts]
      summary: Get posts by user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: User posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Post"
                  hasMore:
                    type: boolean
                required: [posts, hasMore]
  /posts/{postId}:
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: unlikePost
      tags: [posts]
      summary: Unlike a post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /posts/{postId}/comments:
    post:
      operationId: addComment
      tags: [posts]
      summary: Add comment to post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required: [content]
      responses:
        "200":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /posts/{postId}/comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [posts]
      summary: Delete comment
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # STORIES
...
  /reports:
    post:
      operationId: createReport
      tags: [reports]
      summary: Submit a report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "200":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/reports:
    get:
      operationId: getAdminReports
      tags: [reports]
      summary: Get all reports (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, resolved, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: "#/components/schemas/Report"
                required: [reports]
  /admin/reports/{reportId}:
    get:
      operationId: getAdminReport
      tags: [reports]
      summary: Get report detail (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Resolve or reject report (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [resolved, rejected]
              required: [status]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/content/{type}/{id}:
    delete:
      operationId: deleteContent
      tags: [reports]
      summary: Delete reported content (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [post, message, story]
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    UserPublic:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
          nullable: true
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isOnline:
          type: boolean
        lastSeenAt:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, moderator, admin]
      required: [id, displayName, isOnline, role]
...
    Post:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        content:
          type: string
          nullable: true
        mediaUrl:
          type: string
          nullable: true
        visibility:
          type: string
          enum: [public, contacts, private]
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLikedByMe:
          type: boolean
        createdAt:
          type: string
        author:
          $ref: "#/components/schemas/UserPublic"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
      required: [id, userId, visibility, likesCount, commentsCount, isLikedByMe, createdAt, author, comments]
    CreatePostBody:
      type: object
      properties:
        content:
          type: string
        mediaUrl:
          type: string
        visibility:
          type: string
          enum: [public, contacts, private]
          default: public
    Comment:
      type: object
      properties:
        id:
          type: string
        postId:
          type: string
        content:
          type: string
        createdAt:
          type: string
        author:
          $ref: "#/components/schemas/UserPublic"
      required: [id, postId, content, createdAt, author]
...
    CreateReportBody:
      type: object
      properties:
        targetType:
          type: string
          enum: [user, post, message, story]
        targetId:
          type: string
        reason:
          type: string
        details:
          type: string
      required: [targetType, targetId, reason]
    Report:
      type: object
      properties:
        id:
          type: string
        reporterId:
          type: string
        targetType:
          type: string
          enum: [user, post, message, story]
        targetId:
          type: string
        reason:
          type: string
        details:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, resolved, rejected]
        createdAt:
          type: string
        resolvedAt:
          type: string
          nullable: true
      required: [id, reporterId, targetType, targetId, reason, status, createdAt]